css/css-shadow-parts/different-host.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Shadow Parts - Different Host</title>
    <meta href="mailto:fergal@chromium.org" rel="author" title="Fergal Daly">
    <link href="http://www.google.com/" rel="author" title="Google">
    <link href="https://drafts.csswg.org/css-shadow-parts/" rel="help">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/shadow-helper.js"></script>
  </head>
  <body>
    <style>#c-e-other::part(partp) { color: #ff0000; }</style>
    <script>installCustomElement("custom-element", "custom-element-template");</script>
    <template id="custom-element-template">
      <style>span { color: #00ff00; }</style>
      <span id="part" part="partp">This text</span>
    </template>
    The following text should be green:
    <custom-element id="c-e"></custom-element>
    <script type="text/javascript">
      "use strict";
      test(function() {
        var el = getElementByShadowIds(document, ["c-e", "part"]);
        assert_equals(window.getComputedStyle(el).color, "rgb(0, 255, 0)");
      }, "Part is not styled when host is not selected");
    </script>
  </body>
</html>
