LayoutTests/imported/w3c/web-platform-tests/css/css-shadow-parts/different-host-expected.txt

The following text should be green:

PASS Part is not styled when host is not selected
